moving test code from test harness file to individual if __name__ == "__main__"
