+++ trees/avl_tree.py
@@ -680,3 +680,147 @@ class AVLNode:
         height = 1 + max(left_height, right_height)
 
         return is_avl, height
+
+
+if __name__ == "__main__":
+    import time
+    from collections import deque
+    import random
+    import cProfile
+
+    from binary_search_tree import BSTNode
+    from valid_bst import isValidBST
+
+    N = 8 * 10**2
+
+    def is_sorted(arr):
+        return all(arr[i] <= arr[i + 1] for i in range(len(arr) - 1))
+
+    def bst_benchmark():
+        print("\n[ BST BENCHMARK ]")
+        population = range(1, N * 10)
+
+        # values = random.sample(population=population, k=N)
+        values = list(range(N))
+        root = BSTNode(0)
+
+        # Insert
+        t0 = time.time()
+        for v in values:
+            root = BSTNode.insert(root=root, key=v)
+        t1 = time.time()
+        print(f"BST insert {N} values: {t1-t0:.4f}s")
+
+        # In-order traversal
+        def in_order(node: BSTNode):
+            if not node:
+                return []
+            return in_order(node.left) + [node.val] + in_order(node.right)
+
+        t0 = time.time()
+        order = in_order(root)
+        t1 = time.time()
+        print(f"BST in-order traversal: {t1-t0:.4f}s, sorted: {order == sorted(set(order))}")
+
+        # Search
+        t0 = time.time()
+        for v in values:
+            _, _ = root.binary_search(target=v)
+        t1 = time.time()
+        print(f"BST full search: {t1-t0:.4f}s")
+
+        # Delete
+        random.shuffle(values)
+        t0 = time.time()
+        for v in values:
+            root = BSTNode.delete(root, v)
+        t1 = time.time()
+        print(f"BST delete {N} values: {t1-t0:.4f}s")
+        print("BST benchmark complete.\n")
+
+    def avl_bst_benchmark():
+        print("\n[ AVL BENCHMARK ]")
+        population = range(1, N * 10)
+
+        # values = random.sample(population=population, k=N)
+        values = list(range(N))
+        root = AVLNode(0)
+
+        # Insert
+        t0 = time.time()
+        for v in values:
+            root = AVLNode.insert(root=root, key=v)
+        t1 = time.time()
+        print(f"AVL insert {N} values: {t1-t0:.4f}s")
+
+        # In-order traversal
+        def in_order(node: AVLNode):
+            if not node:
+                return []
+            return in_order(node.left) + [node.val] + in_order(node.right)
+
+        t0 = time.time()
+        order = in_order(root)
+        t1 = time.time()
+        print(f"AVL in-order traversal: {t1-t0:.4f}s, sorted: {order == sorted(set(order))}")
+
+        # Search
+        t0 = time.time()
+        for v in values:
+            _, _ = root.binary_search(target=v)
+        t1 = time.time()
+        print(f"AVL full search: {t1-t0:.4f}s")
+
+        # Delete
+        random.shuffle(values)
+        t0 = time.time()
+        for v in values:
+            root = AVLNode.delete(root, v)
+        t1 = time.time()
+        print(f"AVL delete {N} values: {t1-t0:.4f}s")
+        print("AVL benchmark complete.\n")
+
+    def generate_samples() -> tuple[AVLNode, AVLNode]:
+        print("Setting up...")
+        population = range(N * 10)
+        sample = random.sample(population=population, k=N)
+
+        queue = deque(sample)
+        value = queue.popleft()
+        root = AVLNode(value)
+        avl_root = AVLNode(value)
+        assert id(root) != id(avl_root)
+
+        while queue:
+            value = queue.popleft()
+            if value is None:
+                continue
+            root = AVLNode.bst_insert(root=root, key=value)
+            avl_root = AVLNode.insert(root=avl_root, key=value)
+
+        print("ASSERTIONS")
+
+        assert isValidBST(root=root)
+        print("TEST - isValidBST: ✅")
+
+        assert isValidBST(root=avl_root)
+        print("CONTROL - isValidBST: ✅")
+
+        assert not root.is_avl()
+        print("TEST - is_avl: ❌")
+
+        assert avl_root.is_avl()
+        print("CONTROL - is_avl: ✅")
+
+        test_order = list(root.dfs_in_order_traverse())
+        assert is_sorted(test_order)
+        print("TEST - in_order_traverse sorted: ✅")
+
+        control_order = list(avl_root.dfs_in_order_traverse())
+        assert is_sorted(control_order)
+        print("CONTROL - in_order_traverse sorted: ✅")
+
+        return root, avl_root
+
+    # cProfile.run("bst_benchmark()")
+    cProfile.run("avl_bst_benchmark()")
